Fix and expand test_method_missing.
Singular's method_missing recurses infinitely and thus overflows the
stack when it is passed a method the including object doesn't respond
to.

On Ruby 1.8, the resulting SystemStackError is a StandardError, so it
is caught by the rescue statement modifier on line 14 of
lib/structures.rb. But in Ruby 1.9, SystemStackError is a plain 'ol
Exception and thus is not caught.

Prior to this commit, test_method_missing always passed before
method_missing was called because NoMethodError was raised during the
initialization of the Feed object (line 246 of lib/structures.rb) when
parser was called on the nil argument passed in. Fixed this by using a
valid feed object.

Also added an assertion to test the situation in which I discovered the
infinite recursion problem.

test_method_missing should now succeed on 1.8 and fail on 1.9.