Fix infinite recursion problem on Ruby 1.9.
Check whether object responds to pluralized method before sending it to
avoid recursing into method_missing until the stack overflows. The
resulting SystemStackError was caught by the rescue statement modifier
in Ruby 1.8 (with a noticeable pause during the recursion), but isn't
in Ruby 1.9.

Also, remove an assertion that fails now that respond_to? is called.
Could find a way to fix it, but it's redundant.

This fixes test_method_missing on Ruby 1.9; however, there are still 4
unit tests failing on 1.9 for other reasons.